fix warnings

+++ kernel/drivers/terminal_emu.c
@@ -25,7 +25,7 @@ void term_draw_char(char c,terminal_emu_settings *terminal_settings){
 		terminal_settings->y += header->characterSize +1;
 		//some out of bound check
 		if(terminal_settings->y / header->characterSize + 1 >= terminal_settings->height){
-			vfs_ioctl(terminal_settings->frambuffer_dev,IOCTL_FRAMEBUFFER_SCROLL,(void *)header->characterSize + 1);
+			vfs_ioctl(terminal_settings->frambuffer_dev,IOCTL_FRAMEBUFFER_SCROLL,(void *)(uint64_t)header->characterSize + 1);
 			terminal_settings->y -= header->characterSize + 1;
 		}
 		return;
